requirements.txt
torch~=1.8.2+cu111
tqdm==4.56.2
torchvision~=0.9.2+cu111
numpy==1.19.5
Pillow==8.1.0
matplotlib~=3.5.2
